% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_SumStat_sga.R
\name{plot.SumStat_sga}
\alias{plot.SumStat_sga}
\title{Plot the distribution of propensity scores and subgroup balance statistics via Connect-S plot}
\usage{
\method{plot}{SumStat_sga}(x, varlist = NULL, base = FALSE, plotsub = FALSE, ...)
}
\arguments{
\item{x}{a \code{SumStat_sga} object obtained with \code{\link{SumStat}} function.}

\item{varlist}{an optional vector specifying the variables to be plotted on the x-axis by index or variable names.
The default plots all main effect variables specified by \code{ps.formula} in the \code{SumStat_sga} function,
excluding subgrouping variables.}

\item{base}{an indicator to specify whether to print the unadjusted ASMDs before weighting. The default is FALSE.}

\item{plotsub}{an indicator to specify whether to treat subgrouping variables as confounders and print them on the x-axis.
The default is FALSE.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Plot of the indicated type.
}
\description{
Summarize the SumStat_sga object, plot the subgroup balance statistics under weighting versus no weighting.
In a Connect-S plot (Yang et al. 2021), each row represents a subgroup and each column represents a confounder;
each dot is shade-coded according to the value of the ASMD corresponding to the specific subgroup and confounder.
Dark gray and black dots in the Connect-S plot flag meaningful covariate imbalance. The last two columns show the
subgroup sample size and estimated variance inflation due to weighting.
}
\details{
For the Connect-S plot, the shade is categorized based on the common ASMD threshold of 0.1 and 0.2
following Austin and Stuart (2015), with darker shade implying more severe imbalance.
}
\references{
Yang, S., Lorenzi, E., Papadogeorgou, G., Wojdyla, D. M., Li, F., & Thomas, L. E. (2021).
Propensity score weighting for causal subgroup analysis. Statistics in medicine, 40(19), 4294-4309.

Austin, P.C. and Stuart, E.A. (2015). Moving towards best practice when using inverse probability of treatment weighting (IPTW) using the propensity score to estimate causal treatment effects in observational studies.
Statistics in Medicine, 34(28), 3661-3679.
}
