% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psdata_sga.R
\docType{data}
\name{psdata_sga}
\alias{psdata_sga}
\title{Simulated dataset for PSweight.sga}
\format{
A data frame with 3000 rows and 22 columns.
}
\usage{
data(psdata_sga)
}
\description{
This is a simulated observational study with two treatment groups to illustrate the unity of PSweight.sga
}
\details{
The simulated dataset includes 3000 rows, with each row represents information recorded from each individual.
There are 22 variables (columns). The treatment is the variable Treatment, which has two treatment arms.
The outcome of interest is variable Y. X1-X20 are pre-treatment covariates among which X1-X8 are binary, and X9-X20 are continuous.
}
\examples{
data("psdata_sga")
}
\keyword{dataset}
