% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSmethod_sga.R
\name{PSmethod_sga}
\alias{PSmethod_sga}
\title{Fitting propensity scores with different models}
\usage{
PSmethod_sga(
  ps.formula = ps.formula,
  method = "glm",
  weight = "overlap",
  data = data
)
}
\arguments{
\item{ps.formula}{an object of class \code{\link{formula}} (or one that can be coerced to that class):
a symbolic description of the propensity score model to be fitted. Additional details of model specification
are given under "Details". This argument is optional if \code{ps.estimate} is not \code{NULL}.}

\item{method}{a character to specify the method for propensity model. When \code{ps.formula} is given, \code{"glm"} is the default; When \code{xname} is given, \code{"LASSO"} is the default.}

\item{weight}{a character or vector of characters including the types of weights to be used. \code{"IPW"} specifies the inverse probability weights for estimating the average treatment effect among the combined population (ATE). \code{"treated"} specifies the weights for estimating the average treatment effect among the treated (ATT). \code{"overlap"} specifies the (generalized) overlap weights for estimating the average treatment effect among the overlap population (ATO), or population at clinical equipoise. \code{"matching"} specifies the matching weights for estimating the average treatment effect among the matched population (ATM). \code{"entropy"} specifies the entropy weights for the average treatment effect of entropy weighted population (ATEN). Default is \code{"overlap"}.}

\item{data}{a data frame containing the variables in the propensity score model.}
}
\value{
\describe{

\item{\code{ e.h}}{a data frame of estimated propensity scores.}

\item{\code{ nonzero_coef}}{the LASSO selected interactions when \code{method = "LASSO"}.}
}
}
\description{
The function \code{PSmethod_sga} is an internal function to estimate the propensity scores given a specified model through formula.
It is built into function \code{Sumstat_sga}, and \code{PSweight_sga}.
}
\details{
A typical form for \code{ps.formula} is \code{treatment ~ terms} where \code{treatment} is the treatment
variable (identical to the variable name used to specify \code{zname}) and \code{terms} is a series of terms
which specifies a linear predictor for \code{treatment}. \code{ps.formula} by default specifies generalized
linear models given the default argument \code{method = "glm"}.  It fits the logistic regression. The argument
\code{method} allows user to choose model other than glm to fit the propensity score models. In Yang et al.(2021),
the \code{term} is suggested to include all main effects and pairwise subgroup-confounder interactions, combined with
the \code{method}=\code{"LASSO"} to select important interactions. We have included \code{"LASSO"} and \code{"gbm"},
through the \code{method} argument. Note that the current code does not handle multiple treatment groups.
}
\references{
Yang, S., Lorenzi, E., Papadogeorgou, G., Wojdyla, D. M., Li, F., & Thomas, L. E. (2021).
Propensity score weighting for causal subgroup analysis. Statistics in medicine, 40(19), 4294-4309.
}
