% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_PSweight_sga.R
\name{summary.PSweight_sga}
\alias{summary.PSweight_sga}
\title{Summarize a PSweight_sga object}
\usage{
\method{summary}{PSweight_sga}(object, contrast = NULL, type = "DIF", het = FALSE, ...)
}
\arguments{
\item{object}{a PSweight_sga object obtained from the \code{\link{PSweight_sga}} function.}

\item{contrast}{a vector or matrix specifying the causal contrast of interest. The average causal effects will be
defined by such contrats. For multiple treatments, the contrast parameters are explained in Li and Li (2019)
for estimating general causal effects. Default is all pairwise contrasts between any two treatment groups.}

\item{type}{a character specifying the target estimand. The most commonly seen additive estimand is specified
by \code{type = "DIF"}, abbreviated for weighted difference-in-means. This is the usual pairwise average treatment
effects as those defined in Li, Morgan, and Zaslavsky (2018) and Li and Li (2019). For binary (or count outcomes), we also
allow two ratio estimands: causal relative risk (\code{type = "RR"}) and causal odds ratio (\code{type = "OR"}).
Estimates for these two ratio estimands will be reported on the log scale (log relative risk and log
odds ratio) to improve the approximate for asymptotic normality. With binary outcomes, \code{"DIF"} is the same
as the average causal risk difference. Default is "DIF" if left empty.}

\item{het}{an indicator specifying whether to summarize the test of heterogeneity across subgroup levels. The default is FALSE.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of following values:

\describe{
\item{\code{ trtgrp}}{ a character indicating the treatment group, or target population under ATT weights.}

\item{\code{ estimates}}{ a matrix of subgroup point estimates, standard errors and 95% confidence intervals
for contrasts of interest.}


\item{\code{ contrast}}{ a table listing the specified contrasts of interest.}

\item{\code{ group}}{ a table of treatment group labels corresponding to the output point estimates, provided in results
obtained from \code{\link{PSweight_sga}}.}

}
}
\description{
\code{summary.PSweight_sga} is used to summarize the results from \code{\link{PSweight_sga}}.
The output contains the average subgroup causal effects defined by specific contrasts, as well as their
standard error estimates.
}
\details{
For the \code{contrast} argument, one specifies the contrast of interest and thus defines the target estimand
for comparing treatments. For example, if there are two treatment levels: A and B, the contrast A-B
(i.e., E[Y(A)] - E[Y(B)]) can be specified by \code{c(1,-1)}.

For estimating the causal relative risk (\code{type = "RR"}), the contrast is specified at the log scale. For example,
the contrast A-B (specified by \code{c(1,-1)}) implies the estimation of log\{E[Y(A)]\} - log\{E[Y(B)]\}. For estimating the causal odds
ratio, the contrast is specified at the log odds scale. For example, the contrast A-B (specified by \code{c(1,-1)})
implies the estimation of log\{E[Y(A)]/E[1-Y(A)]\} - log\{E[Y(B)]/E[1-Y(B)]\}.

The variance of the contrasts will be estimated by nonparametric bootstrap.

The argument \code{type} takes one of three options: \code{"DIF"}, \code{"RR"}, or \code{"RR"}, with \code{"DIF"} as
the default option. Typically, \code{"RR"} is relavent for binary or count outcomes, and \code{"OR"} is relavent
only for binary outcomes. \code{"DIF"} applies to all types of outcomes.
}
\references{
Yang, S., Lorenzi, E., Papadogeorgou, G., Wojdyla, D. M., Li, F., & Thomas, L. E. (2021).
Propensity score weighting for causal subgroup analysis. Statistics in medicine, 40(19), 4294-4309.

Li, F., Morgan, K. L., Zaslavsky, A. M. (2018).
Balancing covariates via propensity score weighting.
Journal of the American Statistical Association, 113(521), 390-400.
}
