% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumStat_sga.R
\name{SumStat_sga}
\alias{SumStat_sga}
\title{Calculate summary statistics for propensity score weighting in the subgroups}
\usage{
SumStat_sga(
  subgroup = NULL,
  xname = NULL,
  ps.formula = NULL,
  ps.estimate = NULL,
  trtgrp = NULL,
  zname = NULL,
  yname = NULL,
  data = NULL,
  weight = "overlap",
  method = "glm"
)
}
\arguments{
\item{subgroup}{a vector to specify name of subgroup variables by column index or column names}

\item{xname}{an optional character vector specifying the name of the covariates (confounders) in \code{data}. Only continuous and factors are accepted.}

\item{ps.formula}{an object of class \code{\link{formula}} (or one that can be coerced to that class): a symbolic description of the propensity score model to be fitted. Additional details of model specification are given under "Details". This argument is optional if \code{ps.estimate} is not \code{NULL}.}

\item{ps.estimate}{an optional matrix or data frame containing estimated (generalized) propensity scores for each observation. Typically, this is an N by J matrix, where N is the number of observations and J is the total number of treatment levels. Preferably, the column names of this matrix should match the names of treatment level, if column names are missing or there is a mismatch, the column names would be assigned according to the alphabatic order of treatment levels. A vector of propensity score estimates is also allowed in \code{ps.estimate}, in which case a binary treatment is implied and the input is regarded as the propensity to receive the last category of treatment by alphabatic order, unless otherwise stated by \code{trtgrp}.}

\item{trtgrp}{an optional character defining the "treated" population for estimating the average treatment effect among the treated (ATT). Only necessary if \code{weight = "ATT"}. This option can also be used to specify the treatment (in a two-treatment setting) when a vector argument is supplied for \code{ps.estimate}. Default value is the last group in the alphebatic order.}

\item{zname}{an optional character specifying the name of the treatment variable in \code{data}.}

\item{yname}{an optional vector of characters including the names of outcome in \code{data}.}

\item{data}{an optional data frame containing the variables in the propensity score model. If not found in data, the variables are taken from \code{environment(formula)}.}

\item{weight}{a character or vector of characters including the types of weights to be used. \code{"ATE"} specifies the inverse probability weights for estimating the average treatment effect among the combined population. \code{"ATT"} specifies the weights for estimating the average treatment effect among the treated. \code{"ATO"} specifies the (generalized) overlap weights for estimating the average treatment effect among the overlap population, or population at clinical equipoise. Default is \code{"ATO"}.}

\item{method}{a character to specify the method for propensity model. When \code{ps.formula} is given, \code{"glm"} is the default; When \code{cov} is given, \code{"LASSO"} is the default.}
}
\value{
SumStat_sga returns a \code{SumStat_sga} object including a list of the following value:
treatment group, propensity scores, propensity score weights, effective sample sizes,
and balance statistics.

\describe{
\item{\code{ trtgrp}}{a character indicating the treatment group.}

\item{\code{ propensity}}{a data frame of estimated propensity scores.}

\item{\code{ ps.weight}}{a data frame of propensity score weights.}

\item{\code{ ASD}}{ a table including absolute standardized mean differences in the overall sample and subgroups after weighting.}

\item{\code{ ASD_bs}}{ a table including absolute standardized mean differences in the overall sample and subgroups before weighting.}

\item{\code{ vif}}{ a vector indicating the approximated variance inflation in the overall sample and subgroups after weighting, see Yang et al. (2021)}

\item{\code{ nsubg}}{ a vector indicating the subgroup sample sizes.}

\item{\code{ ess}}{a table of effective sample sizes. This serves as a conservative measure to
characterize the variance inflation or precision loss due to weighting, see Yang et al. (2021).}

\item{\code{ subgoup}}{ a vector indicating name of the specified subgroups.}

\item{\code{ method}}{a character indicating the propensity score method used.}

\item{\code{ nonzero_coef}}{a vector indicating the terms selected by LASSO. Only available when \code{method} is \code{LASSO}.}
}
}
\description{
\code{SumStat_sga} is used to generate distributional plots of the estimated propensity scores and balance
diagnostics after propensity score weighting.
}
\details{
A typical form for \code{ps.formula} is \code{treatment ~ terms} where \code{treatment} is the treatment
variable (identical to the variable name used to specify \code{zname}) and \code{terms} is a series of terms
which specifies a linear predictor for \code{treatment}. \code{ps.formula} specifies logistic or more flexible
models for estimating the propensity scores, when \code{ps.estimate} is \code{NULL}.

When comparing two treatments, \code{ps.estimate} can either be a vector or a two-column matrix of estimated
propensity scores. If a vector is supplied, it is assumed to be the propensity scores to receive the treatment, and
the treatment group corresponds to the last group in the alphabetic order, unless otherwise specified by \code{trtgrp}.
In general, \code{ps.estimate} should have column names that indicate the level of the treatment variable,
which should match the levels given in \code{Z}.
If column names are empty or there is a mismatch, the column names will be created following
the alphabetic order of values in \code{Z}, and the rightmost column of \code{ps.estimate} is assumed
to be the treatment group, when estimating ATT. \code{trtgrp} can also be used to specify the treatment
group for estimating ATT.

The argument \code{zname} and/or \code{yname} is required when \code{ps.estimate}
is not \code{NULL}.

Current version of \code{PSweight_sga} allows for three types of propensity score weights used to estimate ATE, ATT and
ATO. These weights are members of larger class of balancing weights defined in Li, Morgan, and Zaslavsky (2018). The overlap weights can also be considered as
a data-driven continuous trimming strategy without specifying trimming rules, see Li, Thomas and Li (2019).
Additional details on balancing weights and generalized overlap weights for multiple treatment groups are provided in
Li and Li (2019).

The variance will be calculated by nonparametric bootstrap, with \code{R} bootstrap
replications. The default of \code{R} is 50.
}
\references{
Yang, S., Lorenzi, E., Papadogeorgou, G., Wojdyla, D. M., Li, F., & Thomas, L. E. (2021).
Propensity score weighting for causal subgroup analysis. Statistics in medicine, 40(19), 4294-4309.

Li, F., Morgan, K. L., Zaslavsky, A. M. (2018).
Balancing covariates via propensity score weighting.
Journal of the American Statistical Association, 113(521), 390-400.

Li, F., Thomas, L. E., Li, F. (2019).
Addressing extreme propensity scores via the overlap weights. American Journal of Epidemiology, 188(1), 250-257.
}
